# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ohlcv {
  id: number;
  symbol: string;
  timeframe: string;
  timestamp: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface LatestOhlcvSummary {
  symbol: string;
  '1m'?: Ohlcv;
  '5m'?: Ohlcv;
  '15m'?: Ohlcv;
  '1h'?: Ohlcv;
}

export interface FundingRate {
  id: number;
  symbol: string;
  timestamp: string;
  fundingRate: number;
  /** @nullable */
  nextFundingTime?: string | null;
}

export interface OpenInterest {
  id: number;
  symbol: string;
  timestamp: string;
  openInterest: number;
  openInterestValue: number;
}

export interface Liquidation {
  id: number;
  symbol: string;
  timestamp: string;
  side: string;
  price: number;
  quantity: number;
  value: number;
}

export interface OrderbookSnapshot {
  id: number;
  symbol: string;
  timestamp: string;
  bids: string;
  asks: string;
  bidAskRatio: number;
  midPrice: number;
  /** @nullable */
  spread?: number | null;
}

export interface CollectorStatusItem {
  name: string;
  /** @nullable */
  lastRun: string | null;
  recordCount: number;
  status: string;
  /** @nullable */
  latestTimestamp: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface CollectorStatus {
  updatedAt: string;
  collectors: CollectorStatusItem[];
}

export interface TableSummary {
  table: string;
  recordCount: number;
  /** @nullable */
  latestTimestamp: string | null;
  /** @nullable */
  oldestTimestamp: string | null;
  /** @nullable */
  dataAgeMinutes?: number | null;
}

export interface OverviewSummary {
  generatedAt: string;
  totalRecords: number;
  tables: TableSummary[];
}

export interface MarketSnapshot {
  symbol: string;
  generatedAt: string;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  priceChange1h?: number | null;
  /** @nullable */
  fundingRate?: number | null;
  /** @nullable */
  openInterest?: number | null;
  /** @nullable */
  openInterestValue?: number | null;
  /** @nullable */
  liquidations1h?: number | null;
  /** @nullable */
  liquidationsValue1h?: number | null;
  /** @nullable */
  bidAskRatio?: number | null;
  /** @nullable */
  volume1h?: number | null;
}

export type ListOhlcvParams = {
symbol?: string;
timeframe?: ListOhlcvTimeframe;
limit?: number;
/**
 * @nullable
 */
since?: string | null;
};

export type ListOhlcvTimeframe = typeof ListOhlcvTimeframe[keyof typeof ListOhlcvTimeframe];


export const ListOhlcvTimeframe = {
  '1m': '1m',
  '5m': '5m',
  '15m': '15m',
  '1h': '1h',
} as const;

export type GetLatestOhlcvParams = {
symbol?: string;
};

export type ListFundingParams = {
symbol?: string;
limit?: number;
/**
 * @nullable
 */
since?: string | null;
};

export type GetLatestFundingParams = {
symbol?: string;
};

export type ListOpenInterestParams = {
symbol?: string;
limit?: number;
/**
 * @nullable
 */
since?: string | null;
};

export type GetLatestOpenInterestParams = {
symbol?: string;
};

export type ListLiquidationsParams = {
symbol?: string;
limit?: number;
/**
 * @nullable
 */
since?: string | null;
};

export type ListOrderbookParams = {
symbol?: string;
limit?: number;
};

export type GetLatestOrderbookParams = {
symbol?: string;
};

export type GetMarketSnapshotParams = {
symbol?: string;
};