# lib/api-zod/src/generated/types/collectorStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectorStatusItem } from './collectorStatusItem';

export interface CollectorStatus {
  updatedAt: string;
  collectors: CollectorStatusItem[];
}