# lib/api-zod/src/generated/types/collectorStatusItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */

export interface CollectorStatusItem {
  name: string;
  /** @nullable */
  lastRun: string | null;
  recordCount: number;
  status: string;
  /** @nullable */
  latestTimestamp: string | null;
  /** @nullable */
  errorMessage?: string | null;
}