# lib/api-zod/src/generated/types/fundingRate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */

export interface FundingRate {
  id: number;
  symbol: string;
  timestamp: string;
  fundingRate: number;
  /** @nullable */
  nextFundingTime?: string | null;
}