# lib/api-zod/src/generated/types/latestOhlcvSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */
import type { Ohlcv } from './ohlcv';

export interface LatestOhlcvSummary {
  symbol: string;
  '1m'?: Ohlcv;
  '5m'?: Ohlcv;
  '15m'?: Ohlcv;
  '1h'?: Ohlcv;
}