# lib/api-zod/src/generated/types/liquidation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */

export interface Liquidation {
  id: number;
  symbol: string;
  timestamp: string;
  side: string;
  price: number;
  quantity: number;
  value: number;
}