# lib/api-zod/src/generated/types/openInterest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenInterest {
  id: number;
  symbol: string;
  timestamp: string;
  openInterest: number;
  openInterestValue: number;
}