# lib/api-zod/src/generated/types/orderbookSnapshot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quant Bot Data Infrastructure API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderbookSnapshot {
  id: number;
  symbol: string;
  timestamp: string;
  bids: string;
  asks: string;
  bidAskRatio: number;
  midPrice: number;
  /** @nullable */
  spread?: number | null;
}